Add files via upload
Color test for the home page.

This was mainly about incorporating some of my favorite element from other artist pages while choosing my favorite colors.
The black one is most similar to the examples I was looking at. I wanted to capture a sort of space like theme and believe black is easy on the eyes so it makes for a more enjoyable process to look at

the blue and purple are varients of my favorite colors and I added a gradient to the background for more visual interest that didn't really work with the black one. 

the major difference is the font choice black and purple are the two i like the most. Black has thinner lines but stylized and the purple has a more painterly type to it